refactor: 일정 생성 퍼널 리펙토링
